requirements.txt
pipecat-ai
fastapi
uvicorn[standard]
websockets
python-dotenv
loguru
google-generativeai
aiohttp
pytz